+++ spec/hatchet/uv_spec.rb
@@ -119,6 +119,7 @@ RSpec.describe 'uv support' do
           PYTHONPATH=/app
           PYTHONUNBUFFERED=true
           WEB_CONCURRENCY=2
+          WEB_CONCURRENCY_SET_BY=heroku/python
         OUTPUT
         expect($CHILD_STATUS.exitstatus).to eq(0)
       end